refactor: middlewares folder structure
