+++ miamoremustachio/ToDo/modules/middlewares/searching/task-searching.js
@@ -1,5 +1,5 @@
-const { tasks } = require('../services/task-services.js');
-const { NotFoundError } = require('../errors/not_found-error.js');
+const { tasks } = require('../../services/task-services.js');
+const { NotFoundError } = require('../../errors/not_found-error.js');
 
 const findTask = async (req, res, next) => {
   const taskId = req.params.id;
